+++ CHANGELOG.md
@@ -47,7 +47,7 @@ The format is based on [Keep a Changelog] and this project adheres to
 
 - Pointer methods for `SEL`.
 
-- Implemented `Default` for `NSString`.
+- Implemented `Default` for `NSString` and `NSMutableString`.
 
 ### Fixed